API

AddLatest.AbstractRegistry

AddLatest.GitRegistry

AddLatest.LocalFolderRegistry

AddLatest.RegistryParsingCache

AddLatest.VersionedPackage

AddLatest.latest

AddLatest.latest

AddLatest.parse_registry

AddLatest.parse_registry

AddLatest.AbstractRegistry — Type

Supertype for all registry types.

AddLatest.GitRegistry — Type

A registry that is cloned using Git.

Examples

julia> registries = [GitRegistry("https://github.com/JuliaRegistries/General.git")]

julia> Pkg.add(latest("Example"; registries = registries))

AddLatest.LocalFolderRegistry — Type

A registry that already exists in a folder on the filesystem.

Examples

julia> registries = [LocalFolderRegistry("/path/to/my/registry/")]

julia> Pkg.add(latest("Example"; registries = registries))

AddLatest.RegistryParsingCache — Type

A cache for holding the results of parsed registries.

Examples

julia> cache = RegistryParsingCache()

julia> Pkg.add(latest("PkgA", cache))

julia> Pkg.add(latest("PkgB", cache))

julia> Pkg.add(latest("PkgC", cache))

AddLatest.VersionedPackage — Type

A specific version of a package.

Fields:

name::String

uuid::UUID

version::String

tree::String

AddLatest.latest — Method

latest(name::AbstractString,
       uuid::UUID;
       cache = RegistryParsingCache(),
       registries = [],
       include_default_registries::Bool = true)

Get the latest registered version of a package, given the name and UUID of the package.

Examples

julia> Pkg.add(latest("Example", UUID("7876af07-990d-54b4-ab0e-23690620f79a")))

AddLatest.latest — Method

latest(name::AbstractString;
       cache = RegistryParsingCache(),
       registries = [],
       include_default_registries::Bool = true)

Get the latest registered version of a package, given the name of the package.

Examples

julia> Pkg.add(latest("Example"))

AddLatest.parse_registry — Method

parse_registry(registry::GitRegistry)

Clone a registry via Git and parse it.

Example usage: ```julia julia> registry = GitRegistry("https://github.com/JuliaRegistries/General.git")

julia> parseresult = parse_registry(registry)

AddLatest.parse_registry — Method

parse_registry(registry::LocalFolderRegistry)

Parse a registry located at the filesystem path path_to_registry.

Example usage:

julia> registry = LocalFolderRegistry("/path/to/my/registry/")
julia> parseresult = parse_registry(registry)